feat: 마이페이지 파일 삭제시 confirm

+++ src/pages/mypage.js
@@ -51,7 +51,10 @@ export default function MyPage() {
       fileType: tab,
     };
     const result = await createFile(newFile);
-    if (result) setUserFiles((prevUserFiles) => [...prevUserFiles, newFile]);
+    console.log(result);
+    if (result) {
+      setUserFiles((prevUserFiles) => [...prevUserFiles, newFile]);
+    }
   };
 
   const renderFileContent = (fileId) => {
@@ -73,9 +76,13 @@ export default function MyPage() {
   };
 
   const handleFileDelete = async (fileId) => {
-    await deleteFile(fileId);
-    setActiveFile(null);
-    setUserFiles((prevFiles) => prevFiles.filter((f) => f.id !== fileId));
+    const confirmResult = window.confirm("파일을 삭제하시겠습니까?");
+
+    if (confirmResult) {
+      await deleteFile(fileId);
+      setActiveFile(null);
+      setUserFiles((prevFiles) => prevFiles.filter((f) => f.id !== fileId));
+    }
   };
 
   return (
